requirements.txt
datasets
pandas==2.1.4
sentence-transformers==2.5.1
faiss-cpu==1.7.4
openai==1.12.0
python-dotenv==1.0.0
graphviz==0.20.1
streamlit==1.44.0
PyMuPDF
matplotlib==3.8.2
Pillow==10.2.0
opencv-python-headless
seaborn==0.13.2
numpy==1.26.3
matplotlib-venn==0.11.10
networkx==3.2.1
nltk==3.9.1
PyPDF2==3.0.1
scikit-learn==1.4.0
torch==2.1.2
transformers==4.37.2